Rename scripts in preperation for repo split
