+++ src/meeting_stats.rb
