+++ src/pipeline_visibility_stats.rb
